@@ -1,6 +1,6 @@
 {
   "name": "@conpagoaus/mongoose-cursor-pagination",
-  "version": "0.3.0",
+  "version": "0.5.0",
   "description": "Easy-to-use, scalable, cursor pagination plugin for mongoose",
   "license": "MIT",
   "homepage": "https://github.com/mother/mongoose-cursor-pagination#readme",
@@ -53,21 +53,22 @@
     "lodash.get": "^4.4.2"
   },
   "devDependencies": {
-    "@types/jest": "^29.2.2",
+    "@types/jest": "^29.2.3",
+    "@types/lodash.get": "^4.4.7",
     "@types/node": "^18.11.9",
-    "eslint": "^8.27.0",
+    "eslint": "^8.28.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-jest": "^27.1.4",
+    "eslint-plugin-jest": "^27.1.6",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.2",
     "jest": "^29.3.1",
     "mongodb-memory-server": "^8.10.0",
-    "mongoose": "^6.7.2",
-    "prettier": "^2.7.1",
+    "mongoose": "^6.7.3",
+    "prettier": "^2.8.0",
     "ts-jest": "^29.0.3",
-    "typescript": "4.8"
+    "typescript": "4.9"
   },
   "peerDependencies": {
-    "mongoose": "^6.7.2"
+    "mongoose": "^6.7.3"
   }
 }
